requirements.txt
# Copyright (C) 2020, 2023 Mitsubishi Electric Research Laboratories (MERL).
#
# SPDX-License-Identifier: AGPL-3.0-or-later

pytorch==1.1.0
gym==0.15.4
tensorboard==2.0.2
numpy==1.18.0
